refactor(android): adopt agent client toolkit
- replace embedded conversational API and transcript code with toolkit 2.9.0
- migrate Living and SIP flows to the published API models
- retain the v1 renderer under the legacy UI package
- raise the minimum SDK to 26 and refresh integration docs
