--- Android/scenes/convoai/src/main/java/io/agora/scene/convoai/convoaiApi/subRender/v1/README.md
@@ -1,3 +1,0 @@
-#  v1 Directory Description
-
-This directory is for demo or internal compatibility use only. Business developers should not directly reference or integrate the contents of this directory. For official integration, please use the v3 directory.
