+++ Android/scenes/convoai/src/main/java/io/agora/scene/convoai/ui/living/CovLivingViewModel.kt
@@ -25,28 +25,28 @@ import io.agora.scene.convoai.api.CovAvatar
 import io.agora.scene.convoai.constant.AgentConnectionState
 import io.agora.scene.convoai.constant.CovAgentManager
 import io.agora.scene.convoai.constant.VoiceprintMode
-import io.agora.scene.convoai.convoaiApi.AgentState
-import io.agora.scene.convoai.convoaiApi.ChatMessageType
-import io.agora.scene.convoai.convoaiApi.ConversationalAIAPIConfig
-import io.agora.scene.convoai.convoaiApi.ConversationalAIAPIError
-import io.agora.scene.convoai.convoaiApi.ConversationalAIAPIImpl
-import io.agora.scene.convoai.convoaiApi.IConversationalAIAPI
-import io.agora.scene.convoai.convoaiApi.IConversationalAIAPIEventHandler
-import io.agora.scene.convoai.convoaiApi.ImageMessage
-import io.agora.scene.convoai.convoaiApi.InterruptEvent
-import io.agora.scene.convoai.convoaiApi.MessageError
-import io.agora.scene.convoai.convoaiApi.MessageReceipt
-import io.agora.scene.convoai.convoaiApi.Metric
-import io.agora.scene.convoai.convoaiApi.ModuleError
-import io.agora.scene.convoai.convoaiApi.ModuleType
-import io.agora.scene.convoai.convoaiApi.Priority
-import io.agora.scene.convoai.convoaiApi.StateChangeEvent
-import io.agora.scene.convoai.convoaiApi.TextMessage
-import io.agora.scene.convoai.convoaiApi.Transcript
-import io.agora.scene.convoai.convoaiApi.TranscriptRenderMode
-import io.agora.scene.convoai.convoaiApi.TranscriptStatus
-import io.agora.scene.convoai.convoaiApi.Turn
-import io.agora.scene.convoai.convoaiApi.VoiceprintStateChangeEvent
+import io.agora.conversational.api.AgentState
+import io.agora.conversational.api.ChatMessageType
+import io.agora.conversational.api.ConversationalAIAPIConfig
+import io.agora.conversational.api.ConversationalAIAPIError
+import io.agora.conversational.api.ConversationalAIAPIImpl
+import io.agora.conversational.api.IConversationalAIAPI
+import io.agora.conversational.api.IConversationalAIAPIEventHandler
+import io.agora.conversational.api.ImageMessage
+import io.agora.conversational.api.InterruptEvent
+import io.agora.conversational.api.MessageError
+import io.agora.conversational.api.MessageReceipt
+import io.agora.conversational.api.Metric
+import io.agora.conversational.api.ModuleError
+import io.agora.conversational.api.ModuleType
+import io.agora.conversational.api.Priority
+import io.agora.conversational.api.StateChangeEvent
+import io.agora.conversational.api.TextMessage
+import io.agora.conversational.api.Transcript
+import io.agora.conversational.api.TranscriptRenderMode
+import io.agora.conversational.api.TranscriptStatus
+import io.agora.conversational.api.Turn
+import io.agora.conversational.api.VoiceprintStateChangeEvent
 import io.agora.scene.convoai.rtc.CovRtcManager
 import io.agora.scene.convoai.rtm.CovRtmManager
 import io.agora.scene.convoai.rtm.IRtmManagerListener
