+++ Android/scenes/convoai/src/main/java/io/agora/scene/convoai/ui/living/messages/CovMessageListView.kt
@@ -21,9 +21,9 @@ import io.agora.scene.common.constant.SSOUserManager
 import io.agora.scene.common.util.GlideImageLoader
 import io.agora.scene.common.util.dp
 import io.agora.scene.convoai.constant.CovAgentManager
-import io.agora.scene.convoai.convoaiApi.Transcript
-import io.agora.scene.convoai.convoaiApi.TranscriptStatus
-import io.agora.scene.convoai.convoaiApi.TranscriptType
+import io.agora.conversational.api.Transcript
+import io.agora.conversational.api.TranscriptStatus
+import io.agora.conversational.api.TranscriptType
 import io.agora.scene.convoai.databinding.CovMessageAgentItemBinding
 import io.agora.scene.convoai.databinding.CovMessageListViewBinding
 import io.agora.scene.convoai.databinding.CovMessageMineItemBinding
