+++ Android/scenes/convoai/src/main/java/io/agora/scene/convoai/ui/living/widget/AgentControlView.kt
@@ -8,7 +8,7 @@ import android.view.LayoutInflater
 import android.widget.LinearLayout
 import io.agora.scene.common.R
 
-import io.agora.scene.convoai.convoaiApi.AgentState
+import io.agora.conversational.api.AgentState
 import io.agora.scene.convoai.databinding.CovWidgetAgentStateBinding
 
 /**
@@ -164,4 +164,4 @@ class AgentControlView @JvmOverloads constructor(
             showNormalState()
         }
     }
-} 
+}
