+++ Android/scenes/convoai/src/main/java/io/agora/scene/convoai/ui/living/widget/AgentStateIndicator.kt
@@ -12,7 +12,7 @@ import android.view.animation.LinearInterpolator
 import androidx.core.content.ContextCompat
 import io.agora.scene.common.R
 import io.agora.scene.common.util.dp
-import io.agora.scene.convoai.convoaiApi.AgentState
+import io.agora.conversational.api.AgentState
 
 /**
  * Agent State Indicator Component
@@ -228,4 +228,4 @@ class AgentStateIndicator @JvmOverloads constructor(
         listeningAnimator?.cancel()
         listeningAnimator = null
     }
-} 
+}
